requirements.txt
asgiref==3.4.1
autopep8==1.5.7
dj-database-url==0.5.0
Django==3.2.6
gunicorn==20.1.0
pycodestyle==2.7.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
Unipath==1.1
whitenoise==5.3.0
django-import-export==2.6.1
google-api-python-client
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
python-dotenv==0.19.0
django-bootstrap-datepicker-plus==3.0.5
social-auth-app-django==5.0.0
social-auth-core==4.1.0
flask==2.0.2
google-auth-oauthlib==0.4.6
google-apis-oauth-django==1.0.0